Fixing type issues (overflows)

+++ libs/minifmod/src/minifmod.cpp
@@ -219,7 +219,7 @@ int FMUSIC_GetRow() noexcept
 */
 int FMUSIC_GetTime() noexcept
 {
-	return FSOUND_last_player_state ? FSOUND_last_player_state->getMixer().getTimeInfo().samples * 1000ull / FSOUND_MixRate : 0;
+	return static_cast<int>(FSOUND_last_player_state ? FSOUND_last_player_state->getMixer().getTimeInfo().samples * 1000ull / FSOUND_MixRate : 0);
 }
 
 bool FSOUND_Init(int mixrate, int /* vcmmode */) noexcept
